+++ Store/InventoryDump.cs
@@ -19,7 +19,7 @@ namespace StoreAndCraft
                 return;
 
             float dumpRange = Plugin.Settings.PlayerDumpRange.Value;
-            NearbyIndex.Rescan(player.transform.position, dumpRange);
+            NearbyIndex.Tick();
             var items = new List<ItemDrop.ItemData>(inv.GetAllItems());
             int stored = 0;
 
@@ -67,7 +67,7 @@ namespace StoreAndCraft
                 return false;
             }
 
-            NearbyIndex.Rescan(player.transform.position, Plugin.Settings.PlayerDumpRange.Value);
+            NearbyIndex.Tick();
             Container chest = ChestPicker.FindStoreTarget(
                 player.transform.position,
                 item,